feat(pool): fix problem with apr tooltip
